refactor: HistoryCellContent 파일 분리
