+++ ProjectManager/ProjectManager/Presentation/TodoHistory/ViewModel/HistoryViewModel.swift
@@ -33,27 +33,3 @@ extension DefaultTodoHistoryViewModel: HistoryViewModel {
             }
     }
 }
-
-struct HistoryCellContent {
-    let title: String
-    let createAt: String
-    
-    private let dateFormatter: DateFormatter = {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy. MM. d"
-        return dateFormatter
-    }()
-    
-    init(item: History) {
-        switch item.changes {
-        case .moved:
-            title = "\(item.changes.rawValue) '\(item.title)' from \(item.beforeState?.rawValue ?? "") to \(item.afterState?.rawValue ?? "")"
-        case .added:
-            title = "\(item.changes.rawValue) '\(item.title)'"
-        case .removed:
-            title = "\(item.changes.rawValue) '\(item.title)' from \(item.beforeState?.rawValue ?? "")"
-        }
-        
-        self.createAt = item.createAt.toString(dateFormatter)
-    }
-}
